addHabit method to viewmodel + create ViewModifiers directory
